requirements.txt
flask
langchain
langchain-openai
langchain-community
qdrant-client
google-generativeai
tiktoken
